Move all deprecated funcs etc... into old_ files, for easy deleting once we get to a major version bump
